added search function to ProductsManage component

+++ src/components/Tabs/ProductsManage/ProductsManage.tsx
@@ -1,6 +1,8 @@
+import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { BsFillTrashFill } from 'react-icons/bs';
+import { BsFillTrashFill, BsSearch } from 'react-icons/bs';
 import { AiFillEdit } from 'react-icons/ai';
+
 import { toast } from 'react-hot-toast';
 import {
   useGetUserProductQuery,
@@ -8,6 +10,8 @@ import {
 } from '../../../store/api/productsApiSlice';
 import Spinner from '../../Spinner/Spinner/Spinner';
 import classes from './ProductsManage.module.scss';
+import Input from '../../UI/Input/Input';
+import { ProductType } from '../../../types/types';
 
 interface PropsType {
   token: string;
@@ -16,8 +20,14 @@ const ProductsManage = ({ token }: PropsType) => {
   const { data: userProducts, isLoading } = useGetUserProductQuery(token, {
     refetchOnMountOrArgChange: true,
   });
+  const [searchValue, setSearchValue] = useState('');
+  const [searchResults, setSearchResults] = useState<ProductType[]>([]);
+
   const [deleteProduct, { isError }] = useDeleteProductMutation();
   let content;
+  const searchValueHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchValue(e.target.value);
+  };
 
   const deleteProductHandler = async (id: string) => {
     // eslint-disable-next-line no-alert
@@ -39,44 +49,62 @@ const ProductsManage = ({ token }: PropsType) => {
       }
     }
   };
+
+  const searchProductsHanlder = () => {
+    if (userProducts) {
+      const filteredProducts = userProducts.products.filter((product) =>
+        product.name.toLowerCase().startsWith(searchValue.toLowerCase())
+      );
+
+      setSearchResults(filteredProducts);
+    }
+  };
+
   if (isLoading) {
     content = <Spinner message="Ładowanie produktów" />;
   } else if (userProducts?.products) {
-    content = (
-      <>
-        {userProducts.products.map((product) => (
-          <div key={product._id} className={classes.productsBox__product}>
-            <div className={classes[`productsBox__product--name`]}>
-              <img src={product.imageUrl} alt={product.name} width={100} />
-              <p>{product.name}</p>
-            </div>
-            <div className={classes[`productsBox__product--buttons`]}>
-              <div>
-                <Link to={`/produkt/${product._id}/edit`}>
+    const productsToDisplay =
+      searchResults.length > 0 ? searchResults : userProducts.products;
+
+    if (productsToDisplay.length === 0) {
+      content = <p>Brak wyników dla podanej frazy wyszukiwania.</p>;
+    } else {
+      content = (
+        <>
+          {productsToDisplay.map((product) => (
+            <div key={product._id} className={classes.productsBox__product}>
+              <div className={classes[`productsBox__product--name`]}>
+                <img src={product.imageUrl} alt={product.name} width={100} />
+                <p>{product.name}</p>
+              </div>
+              <div className={classes[`productsBox__product--buttons`]}>
+                <div>
+                  <Link to={`/produkt/${product._id}/edit`}>
+                    <button
+                      type="button"
+                      className={classes[`productsBox__product--button`]}
+                    >
+                      <AiFillEdit />
+                    </button>
+                  </Link>
+                </div>
+                <div>
                   <button
-                    type="button"
+                    type="submit"
+                    onClick={() => {
+                      deleteProductHandler(product._id);
+                    }}
                     className={classes[`productsBox__product--button`]}
                   >
-                    <AiFillEdit />
+                    <BsFillTrashFill />
                   </button>
-                </Link>
-              </div>
-              <div>
-                <button
-                  type="submit"
-                  onClick={() => {
-                    deleteProductHandler(product._id);
-                  }}
-                  className={classes[`productsBox__product--button`]}
-                >
-                  <BsFillTrashFill />
-                </button>
+                </div>
               </div>
             </div>
-          </div>
-        ))}
-      </>
-    );
+          ))}
+        </>
+      );
+    }
   }
   return (
     <div className={classes.mainContainer}>
@@ -89,6 +117,26 @@ const ProductsManage = ({ token }: PropsType) => {
       </div>
       <div className={classes.productsBox}>
         <h5>Twoje produkty:</h5>
+        <div className={classes.productsBox__search}>
+          <Input
+            type="text"
+            text="Szukaj..."
+            data="search"
+            onChange={searchValueHandler}
+          />
+          <button
+            type="button"
+            onClick={searchProductsHanlder}
+            onKeyDown={(event) => {
+              if (event.key === 'Enter' || event.key === ' ') {
+                searchProductsHanlder();
+              }
+            }}
+            tabIndex={0}
+          >
+            <BsSearch />
+          </button>
+        </div>
         {content}
       </div>
     </div>
